added more tests for searchParams
